requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
charset-normalizer==3.0.1
click==8.1.3
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
decorator==5.1.1
Django==4.1.3
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.2.1
docopt==0.6.2
dynaconf==3.1.12
exceptiongroup==1.1.0
hvac==1.0.2
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
iniconfig==2.0.0
Jinja2==3.1.2
knox==0.1.14
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.2.0
pipreqs==0.4.11
pluggy==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.6.0
pymongo==4.3.3
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pytest==7.2.1
pytz==2022.2.1
requests==2.28.2
service-identity==21.1.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.2.4
tomli==2.0.1
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.14
validators==0.20.0
yarg==0.1.9
zope.interface==5.4.0
